fix sonar on register page
