+++ src/pages/register/register-from/index.js
@@ -1,0 +1,3 @@
+import RegisterForm from './register-form';
+
+export default RegisterForm;
